# docs/swagger.yaml
basePath: /
definitions:
...
  requests.CreateChatRequest:
    properties:
      initial_message:
        description: 最初のメッセージを同時に送信する場合
        example: こんにちは！
        type: string
      partner_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
    required:
    - partner_id
    type: object
...
  requests.SendAvatarChatMessageRequest:
    properties:
      content:
        example: Hello, how are you?
        type: string
    required:
    - content
    type: object
...
  requests.SendMessageRequest:
    properties:
      content:
        example: こんにちは！今日はいい天気ですね。
        type: string
    required:
    - content
    type: object
...
  response.AvatarChatMessage:
    properties:
      created_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      message:
        example: Hello, how are you?
        type: string
      sender_type:
        example: user
        type: string
    type: object
...
  response.Chat:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      is_matched:
        example: false
        type: boolean
      last_message:
        example: こんにちは！
        type: string
      last_message_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      matching_score:
        example: 75
        type: integer
      partner_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
      partner_image_url:
        example: https://example.com/images/profile2.jpg
        type: string
      partner_name:
        example: 佐藤花子
        type: string
      unread_count:
        example: 0
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
    type: object
  response.ChatScore:
    properties:
      chat_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      current_score:
        example: 75
        type: integer
      max_score:
        example: 100
        type: integer
      score_breakdown:
        additionalProperties:
          type: integer
        type: object
      unlock_threshold:
        description: マッチング成立に必要なスコア
        example: 80
        type: integer
      updated_at:
        example: "2024-01-01T12:00:00Z"
        type: string
    type: object
  response.CreateChatResponse:
    properties:
      chat:
        $ref: '#/definitions/response.Chat'
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        example: invalid_token
        type: string
      message:
        example: Firebase IDトークンが不正または無効です
        type: string
    type: object
...
  response.GetAvatarChatMessagesResponse:
    properties:
      is_matched:
        example: false
        type: boolean
      matching_point:
        example: 50
        type: integer
      messages:
        items:
          $ref: '#/definitions/response.AvatarChatMessage'
        type: array
    type: object
  response.GetAvatarChatStatusResponse:
    properties:
      is_matched:
        example: false
        type: boolean
      matching_point:
        example: 50
        type: integer
      unlocked_user_infos:
        items:
          $ref: '#/definitions/response.UnlockedUserInfo'
        type: array
    type: object
  response.GetChatDetailResponse:
    properties:
      chat:
        $ref: '#/definitions/response.Chat'
    type: object
  response.GetChatScoreResponse:
    properties:
      score:
        $ref: '#/definitions/response.ChatScore'
    type: object
  response.GetChatsResponse:
    properties:
      chats:
        items:
          $ref: '#/definitions/response.Chat'
        type: array
      total:
        example: 10
        type: integer
    type: object
...
  response.Message:
    properties:
      chat_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      content:
        example: こんにちは！今日はいい天気ですね。
        type: string
      created_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
      sender_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      sender_type:
        description: '"user" or "avatar_ai"'
        example: user
        type: string
    type: object
...
  response.SendAvatarChatMessageResponse:
    properties:
      avatar_response:
        $ref: '#/definitions/response.AvatarChatMessage'
      is_matched:
        example: false
        type: boolean
      matching_point:
        example: 50
        type: integer
      message:
        example: Message sent successfully
        type: string
      point_change:
        example: 10
        type: integer
      unlocked_missions:
        items:
          $ref: '#/definitions/response.UnlockedUserInfo'
        type: array
    type: object
...
  response.SendMessageResponse:
    properties:
      ai_response:
        allOf:
        - $ref: '#/definitions/response.Message'
        description: AIからの自動返信がある場合
      matching_score:
        example: 75
        type: integer
      message:
        $ref: '#/definitions/response.Message'
    type: object
...
  response.UnlockedUserInfo:
    properties:
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      key:
        example: hobby
        type: string
      value:
        example: reading
        type: string
    type: object
...
  response.User:
    properties:
      age:
        example: 25
        type: integer
      bio:
        example: よろしくお願いします！
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      display_name:
        example: 山田太郎
        type: string
      gender:
        example: male
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      onboarding_completed:
        example: false
        type: boolean
      profile_image_url:
        example: https://example.com/images/profile.jpg
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
  title: Hackathon API
  version: "1.0"
paths:
  /auth/me:
    get:
      description: AuthorizationヘッダーのFirebase IDトークンからログイン中のユーザー情報を取得する
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.User'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: ログイン中ユーザー取得
      tags:
      - auth
  /avatar-chats/{avatar_id}/messages:
    get:
      description: アバターとのチャット履歴を取得する
      parameters:
      - description: アバターID
        in: path
        name: avatar_id
        required: true
        type: string
      responses:
        "200":
          description: チャット履歴取得成功
          schema:
            $ref: '#/definitions/response.GetAvatarChatMessagesResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: アバターチャットメッセージ取得
      tags:
      - avatar-chat
    post:
      description: アバターにメッセージを送信し、応答を受け取る
      parameters:
      - description: アバターID
        in: path
        name: avatar_id
        required: true
        type: string
      - description: メッセージ送信リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SendAvatarChatMessageRequest'
      responses:
        "200":
          description: メッセージ送信成功
          schema:
            $ref: '#/definitions/response.SendAvatarChatMessageResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: アバターチャットメッセージ送信
      tags:
      - avatar-chat
  /avatar-chats/{avatar_id}/status:
    get:
      description: アバターとのマッチングステータスを取得する
      parameters:
      - description: アバターID
        in: path
        name: avatar_id
        required: true
        type: string
      responses:
        "200":
          description: ステータス取得成功
          schema:
            $ref: '#/definitions/response.GetAvatarChatStatusResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: アバターチャットステータス取得
      tags:
      - avatar-chat
...
  /chats:
    get:
      description: 自分が参加しているチャットの一覧を取得する
      responses:
        "200":
          description: チャット一覧取得成功
          schema:
            $ref: '#/definitions/response.GetChatsResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 自分のチャット一覧取得
      tags:
      - chats
    post:
      description: 相手との新しいチャットを開始する
      parameters:
      - description: チャット開始リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateChatRequest'
      responses:
        "200":
          description: チャット開始成功
          schema:
            $ref: '#/definitions/response.CreateChatResponse'
        "400":
          description: リクエストが不正、または既にチャットが存在する
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 相手が見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 新しいチャット開始
      tags:
      - chats
  /chats/{partnerUserId}:
    get:
      description: 特定のチャットの詳細情報を取得する（IDは相手のUserID）
      parameters:
      - description: 相手のユーザーID（ULID）
        in: path
        name: partnerUserId
        required: true
        type: string
      responses:
        "200":
          description: チャット詳細取得成功
          schema:
            $ref: '#/definitions/response.GetChatDetailResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このチャットにアクセスする権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: チャットが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 特定チャットの詳細取得
      tags:
      - chats
...
  /chats/{partnerUserId}/messages:
    post:
      description: 相手AIにメッセージを送信する（IDは相手のUserID、メッセージ履歴はFirestoreから直接取得）
      parameters:
      - description: 相手のユーザーID（ULID）
        in: path
        name: partnerUserId
        required: true
        type: string
      - description: メッセージ送信リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SendMessageRequest'
      responses:
        "200":
          description: メッセージ送信成功
          schema:
            $ref: '#/definitions/response.SendMessageResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このチャットにアクセスする権限がない、またはマッチング済み
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: チャットが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: メッセージ送信
      tags:
      - chats
  /chats/{partnerUserId}/score:
    get:
      description: 特定チャットの現在のマッチングポイントを取得する（IDは相手のUserID）
      parameters:
      - description: 相手のユーザーID（ULID）
        in: path
        name: partnerUserId
        required: true
        type: string
      responses:
        "200":
          description: マッチングポイント取得成功
          schema:
            $ref: '#/definitions/response.GetChatScoreResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このチャットにアクセスする権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: チャットが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 現在のマッチングポイント取得
      tags:
      - chats
...
securityDefinitions:
  Bearer:
    description: 'Firebase IDトークンをBearerトークンとして送信してください。例: "Bearer eyJhbGciOiJSUzI1NiIs..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"